+++ Components/PopularCourses.razor
@@ -1,5 +1,5 @@
 @using System.Globalization
-
+@inject NavigationManager NavigationManager
 <!-- =======================
 Popular course START -->
 <section>
@@ -7,8 +7,8 @@ Popular course START -->
 		<!-- Title -->
 		<div class="row mb-4">
 			<div class="col-lg-8 mx-auto text-center">
-				<h2 class="fs-1">Ko'p so'raladigan kurslar</h2>
-				<p class="mb-0">Sifati kafolatlanga kurslarimizdan o'zingizga yoqqanini tanlang.</p>
+				<h2 class="fs-1">Kurslar ro'yxati</h2>
+				<p class="mb-0">Sifati kafolatlangan kurslarimizdan o'zingizga yoqqanini tanlang.</p>
 			</div>
 		</div>
 
@@ -49,8 +49,8 @@ Popular course START -->
                         @foreach(var course in SelectedCourses)
                         {
                             var score = Math.Round(course.Score * 2, MidpointRounding.AwayFromZero) / 2;
-                            <!-- Card item START -->
                             <div class="col-sm-6 col-lg-4 col-xl-3">
+                            <a href="javascript:void(0)" @onclick="() => SendCourseData(course.Title)" > 
                                 <div class="card shadow h-100">
                                     <!-- Image -->
                                     <img src="assets/images/courses/@course.BannerId" class="card-img-top" alt="@course.Title">
@@ -58,11 +58,10 @@ Popular course START -->
                                     <div class="card-body pb-0">
                                         <!-- Badge and favorite -->
                                         <div class="d-flex justify-content-between mb-2">
-                                            <a href="#" class="badge bg-purple bg-opacity-10 text-purple">@Capitalize(course.Tag)</a>
-                                            <a href="#" class="h6 mb-0"><i class="far fa-heart"></i></a>
+                                            <a class="badge bg-purple bg-opacity-10 text-purple">@Capitalize(course.Tag)</a>
                                         </div>
                                         <!-- Title -->
-                                        <h5 class="card-title fw-normal"><a href="#">@course.Title</a></h5>
+                                        <h5 class="card-title fw-normal"><a >@course.Title</a></h5>
                                         <p class="mb-2 text-truncate-2">@course.Description</p>
                                         <!-- Rating star -->
                                         <ul class="list-inline mb-0">
@@ -99,6 +98,7 @@ Popular course START -->
                                         </div>
                                     </div>
                                 </div>
+                            </a>
                             </div>
                             <!-- Card item END -->
                         }
@@ -111,3 +111,10 @@ Popular course START -->
 </section>
 <!-- =======================
 Popular course END -->
+@code{
+
+    private void SendCourseData(string title)
+    {
+        NavigationManager.NavigateTo($"coursedetail/{title}");
+    }
+}
