+++ Pages/Index.razor
@@ -1,8 +1,6 @@
 ﻿@page "/"
 @using landing.Components.Figures
 
-<!-- =======================
-Main Banner START -->
 <section class="position-relative overflow-hidden pt-5 pt-lg-3">
 	
 	<SvgFigures/>
@@ -12,7 +10,7 @@ Main Banner START -->
 		<!-- Title -->
 		<div class="row align-items-center g-5">
 			<!-- Left content START -->
-			<div class="col-lg-5 col-xl-6 position-relative z-index-1 text-center text-lg-start mb-5 mb-sm-0 border-0">
+			<div class="col-lg-5 col-xl-6 position-relative z-index-1 text-center text-lg-start mb-9 mb-sm-0 border-0">
 				<!-- SVG -->
 				<HeaderFigures/>
 				
@@ -37,11 +35,11 @@ Main Banner START -->
 					<li class="list-inline-item"> <i class="bi bi-patch-check-fill h6 me-1"></i>Hayotiy proyektlar</li>
 				</ul>
 
-				<div class="d-sm-flex align-items-center justify-content-center justify-content-lg-start mb-6">
+				<div class="d-sm-flex align-items-center justify-content-center justify-content-lg-start mb-0">
 					<!-- Button -->
-					<a href="#" class="btn btn-lg btn-danger-soft me-2 mb-sm-0 mb-4">Boshladik</a>
+					@* <a href="#" class="btn btn-lg btn-danger-soft me-2 mb-sm-0 mb-4">Boshladik</a> *@
 					<!-- Video button -->
-          <a data-glightbox data-gallery="office-tour" href="https://www.youtube.com/embed/tXHviS-4ygo" class="ms-0 ms-sm-4 d-block">
+          			<a data-glightbox data-gallery="office-tour" href="https://www.youtube.com/embed/tXHviS-4ygo" class="me-2 mb-sm-0 mb-7 d-block">
 						<div class="btn btn-round btn-primary-shadow mb-0 me-3 align-middle d-inline-block"> <i class="fas fa-play"></i></div>
 						<div class="align-middle d-inline-block">
 							<h6 class="mb-0 fw-normal">Videoni ko'ring</h6>
@@ -101,7 +99,7 @@ Counter START -->
 							<h5 class="purecounter mb-0 fw-bold" data-purecounter-start="0" data-purecounter-end="60"	data-purecounter-delay="200">100</h5>
 							<span class="mb-0 h5">+</span>
 						</div>
-						<p class="mb-0">Muvoffaqiyatli talablar</p>
+						<p class="mb-0">Muvaffaqiyatli talablar</p>
 					</div>
 				</div>
 			</div>
