+++ Pages/Register.razor
@@ -1,0 +1,3 @@
+@page "/register"
+
+<p>register urself</p>
