+++ Shared/Header.razor
@@ -25,8 +25,8 @@
 				<!-- Nav category menu START -->
 				<ul class="navbar-nav navbar-nav-scroll mx-5">
 					<!-- Nav item 1 Demos -->
-					<li class="nav-item dropdown dropdown-menu-shadow-stacked">
-						<a class="nav-link bg-primary bg-opacity-10 rounded-3 text-primary px-3 py-3 py-xl-0" href="#" id="categoryMenu" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false"><i class="bi bi-ui-radios-grid me-2"></i><span>Ro'yhatdan o'tish</span></a>
+					<li class="nav-item">
+						<a class="nav-link bg-primary bg-opacity-10 rounded-3 text-primary px-3 py-3 py-xl-0" href="/register" aria-haspopup="true" aria-expanded="false"><i class="bi bi-ui-radios-grid me-2"></i><span>Ro'yxatdan o'tish</span></a>
 					</li>
 				</ul>
 				<!-- Nav category menu END -->
