+++ _Imports.razor
@@ -9,3 +9,4 @@
 @using landing
 @using landing.Shared
 @using landing.Components
+@using landing.Models
