+++ wwwroot/credential.json
@@ -1,0 +1,12 @@
+{
+  "type": "service_account",
+  "project_id": "robotic-casing-340204",
+  "private_key_id": "c1bd6f5bcc3fa74fcb2d385ed73c75a7b3723487",
+  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQC7R6MPv8VeA0e5\ncx8gh90MCQUqkh9QUhE96PY9e/7yz6vCgLWKXpkSsBc3nwE0/GTIAN6IFfqZZaVn\nOCx4/3yCh6UhcwZJJZGf22YOX3HJbSv6JhWpTK4PdmchgkymYhOPUGJ6JRvVVVTm\nHusPPwaqJcedrrq7FrI5jVk5UbQZvEy6/xJrAUzIryHZLGIwyQPzLZY6tT0hsYH8\nshvdGNRR0HEc+M3uzKAvHOSlOsVnGTNO5zQqRUfZ17kRGLKFTwulqJHbB4tHg8FJ\nP0RK3+yJAg6jaeJoPmY6qNnttc7gbrQVwNbC0jjSxk2GQ6E2O23JLS63uHDef8BR\n6RQOwr5JAgMBAAECggEAEkzZvjw7CQjBBXoeXItqKjbnblWBAdRDIw2UokVvkEKC\nwirZkn/T7WOPoTvEXm0bW//ZvuqG5l5sXz25+sjcXdTrae1Xe+ZT6I8D1fu1nr0w\nQu1dyIFZsOEdCsU0wD1tM9/i/5SIt7vy49m5lm/OMuomK15bmNSK7dzGDzfQpYL5\n6YjMzUkziUHHz7PDOs1L5qW9GS3QYczxVMGc4NrCO4aFfegyd1AP58ezoBSYYqmw\nvcCMAce6Ui0YurYzXFeALx/5plZZwCTRNozK5Pp8lHA6zsUvAppemJOJpDZ+3KaN\nPxAZxadcvc0s6+6QuLGW1MpuQEJP4E1KxoOuLZo4QQKBgQDyRGJLrrQynP1lMkEJ\nnJFNqhy1QJC06TDcf+1EwXTkEfz5kXgrk7kDhlZ9REqx7BqbZeAB0KHT3GhlR3zk\n22G2NeYzfZ7ty2F1iBy4D4fw0JZMmM7UWebnbBSNnqz6wPD7xGTcxBRlqyfCNNW8\nrR8m8p399xLuArvAk2zO7ujr7QKBgQDF5VCSC4gSjp6qhigmlm9ovHK/cP3vDUbM\nHYK8d0cSskjK/uH/mNtbvwRPFSLn23283pVHC0xNsp/7DPgkjDsvHZBy5kh4xKA+\nMC9xI02v+IIskykiiCFmtEU4Auqk/JOM65vnN3XvFzud+2xtcI2qM2ksKvUi0F2z\nKOF1hdnoTQKBgQCEeCXzUhzQDA12Vws+8TB2cgzugtZdRaDlAMJBNTLwPnHAGO3H\njie9KKiEKGJbdUl7LTJWa7kA+mTO9E7zaI2+LyynfQn45ptsw7dIYZ2obLWyPl/h\ntyLN6cg3Q8s79WIPmGy6VjD9AqBjzbYNS/BbwI/pPT9Ji48UFsiJQhN0ZQKBgHQX\neFKUrTilrfK/TOKGMckxc+dWs8D++t9XGJDTMECJkiZzq8SzuUtzHWX/MkDuBC8+\nLDhHKnYSKcOu8gsOsNGM+GZPLYibYJ3aypoGadanZEYd9m2pil6jtk+LdndfLHfo\n58aWRHsEGBcvnYX+Fz0Xd8VPVgscTTuC6fB1m0WpAoGBANb1If5C4pIPQ5vglQ+H\nAMPWK5g0PAwJKbJY28iInS2eKsFvbNhk0PAmbGkjvZZxC6eIXMse3J5UyeqWdGS0\nTagDiXxbuvk59DMp+xWKuod+xcedvWX8KncHFKLmUrJkWfpvU11kPQCCd/Hzhs1j\nbRvTA/GBv98GjQAxq9t2rOgv\n-----END PRIVATE KEY-----\n",
+  "client_email": "ilmhub@robotic-casing-340204.iam.gserviceaccount.com",
+  "client_id": "107078791359990949642",
+  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
+  "token_uri": "https://oauth2.googleapis.com/token",
+  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
+  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/ilmhub%40robotic-casing-340204.iam.gserviceaccount.com"
+}
